Updated 22.3 integer values to decimal

+++ src/EncompassRest/Loans/Correspondent.cs
@@ -56,7 +56,7 @@ namespace EncompassRest.Loans
         private DirtyValue<string?>? _appraisalUsage;
         private DirtyValue<bool?>? _appraisalWaiver;
         private DirtyValue<string?>? _appraisedCondition;
-        private DirtyValue<int?>? _appraisedValue;
+        private DirtyValue<decimal?>? _appraisedValue;
         private DirtyValue<string?>? _appraiserName;
         private DirtyValue<string?>? _appraisersLicenseNumber;
         private DirtyValue<string?>? _appraisersPropertyAddress1;
@@ -750,7 +750,7 @@ namespace EncompassRest.Loans
         /// <summary>
         /// Correspondent Appraised Value [CORRESPONDENT.X96]
         /// </summary>
-        public int? AppraisedValue { get => _appraisedValue; set => SetField(ref _appraisedValue, value); }
+        public decimal? AppraisedValue { get => _appraisedValue; set => SetField(ref _appraisedValue, value); }
 
         /// <summary>
         /// Correspondent Appraiser Name [CORRESPONDENT.X97]
